exposing failing Discard tests
