+++ core/src/test/scala/MonitoringSpec.scala
@@ -554,169 +554,4 @@ object MonitoringSpec extends Properties("monitoring") {
     }
     go || go || go
   }
-
-  /*
-  property("keys are being removed upon end of data from host") = secure {
-    val M = Monitoring.default
-    val i = new Instruments(monitoring = M)
-    val gauge = i.gauge("someGauge",0.0)
-
-    M.keySenescence(Events.every(1 milliseconds),Process.eval(Task.delay(gauge.key))).run.run
-    //TODO: Must refactor to allow simulation of stopped data
-    M.keys.get.run.contains(gauge.key)
-  }
-  */
-
-  property("keys are being removed upon disconnect") = secure {
-    val M = Monitoring.default
-    val i = new Instruments(monitoring = M)
-    val gauge = i.gauge("someGauge",0.0)
-    val key = gauge.key
-
-    //closing the signal, simulating the host dropping the connection
-    M.get(key).close.run
-
-    val x = M.keySenescence(Events.every(1 milliseconds),Process.eval(Task.delay(gauge.key)))
-
-    Thread.sleep(1000)
-
-    x.run.run
-    !M.keys.get.run.contains(key)
-  }
-
-
-  property("TCP disconnect removes host from /mirror/sources list") = secure {
-    val M = Monitoring.default
-    val enqueueSink: Sink[Task, Command] = M.mirroringQueue.enqueue
-    val uri= URI.create("http://localhost")
-    val clusterName = "clusterName"
-    val mirror = Mirror(uri,clusterName)
-    val discard = Discard(uri)
-    val datapoint = new Datapoint[Any](Key[String]("key",Units.TrafficLight),"green")
-    val commands: Process[Task, Command] = Process.emitAll(Seq(mirror, discard))
-
-    //this should be commands.to(enqueueSink) but scala 2.10 didn't like it
-    //val commandEnqueue = commands.to(enqueueSink)
-    val commandEnqueue = commands.zipWith(enqueueSink)((o,f) => f(o)).eval
-
-    val mockParse: URI => Process[Task, Datapoint[Any]] = 
-      _ => Process.eval(Task.delay(throw new RuntimeException("boom")))
-
-    //enqueue the commands
-    commandEnqueue.run.run
-
-    //This is used as a flag to cancel the processing
-    val b = new java.util.concurrent.atomic.AtomicBoolean(false)
-
-    //start processing commands
-    M.processMirroringEvents(
-      mockParse,
-      nodeRetries = _ => Events.takeEvery(1 millisecond,1)
-    ).runAsyncInterruptibly(_ => (), b)
-
-    //Let it run
-    Thread.sleep(1000)
-
-    //end processing
-    b.set(true)
-
-    //cleanup
-    M.mirroringQueue.close.run
-
-    M.mirroringUrls.size == 0
-  }
-
-  private val clusterName: ClusterName = "clusterName"
-  property("Disconnect command removes host from /mirror/sources list") = secure {
-    val M = Monitoring.default
-    val enqueueSink: Sink[Task, Command] = M.mirroringQueue.enqueue
-    val uri= URI.create("http://localhost")
-    val mirror = Mirror(uri,clusterName)
-    val discard = Discard(uri)
-    val datapoint = new Datapoint[Any](Key[String]("key",Units.TrafficLight),"green")
-    val commands: Process[Task, Command] = Process.emitAll(Seq(mirror, discard))
-    val commandEnqueue = commands.zipWith(enqueueSink)((o,f) => f(o)).eval
-    val mockParse: URI => Process[Task, Datapoint[Any]] = _ => scalaz.stream.io.resource(Task.delay(())
-    ){ _ => Task.delay{ () }
-    }(_ => Task.delay(datapoint))
-
-    //enqueue the commands
-    commandEnqueue.run.run
-
-    //This is used as a flag to cancel the processing
-    val b = new java.util.concurrent.atomic.AtomicBoolean(false)
-
-    //start processing commands
-    M.processMirroringEvents(mockParse).runAsyncInterruptibly(_ => (), b)
-
-    Thread.sleep(1000)
-
-    //end processing
-    b.set(true)
-    Thread.sleep(1000)
-
-    //cleanup
-    M.mirroringQueue.close.run
-
-    println(M.mirroringUrls.toMap.get(clusterName))
-    M.mirroringUrls.toMap.get(clusterName) == None
-  }
-
-  /*
-  Test fails due to io.resource's cleanup code not being executed.
-  property("Disconnect Command disconnects from Host")= secure {
-    val M = Monitoring.default
-    val enqueueSink: Sink[Task, Command] = M.mirroringQueue.enqueue
-    val uri= URI.create("http://localhost")
-    val mirror = Mirror(uri,clusterName)
-    val discard = Discard(uri)
-    val datapoint = new Datapoint[Any](Key[String]("key",Units.TrafficLight),"green")
-    val result = new java.util.concurrent.atomic.AtomicBoolean(false)
-    val countdown = new java.util.concurrent.CountDownLatch(1)
-    val commands: Process[Task, Command] = Process.emitAll(Seq(mirror, discard))
-    val commandEnqueue = commands.zipWith(enqueueSink)((o,f) => f(o)).eval
-
-    val mockDataConnection: URI => Process[Task, Datapoint[Any]] = _ => scalaz.stream.io.resource(Task.delay(())){ _ =>
-      Task.delay{
-        result.set(true)
-        countdown.countDown
-        ()
-      }
-
-    }(_ => Task.delay(datapoint))
-
-    //enqueue the commands
-    commandEnqueue.run.run
-
-    //This is used as a flag to cancel the processing
-    val b = new java.util.concurrent.atomic.AtomicBoolean(false)
-
-    //start processing commands
-    M.processMirroringEvents(mockDataConnection).runAsyncInterruptibly(_ => (), b)
-
-    Thread.sleep(1000)
-
-    //check whether the io cleanup code was run
-    countdown.await(3, java.util.concurrent.TimeUnit.SECONDS)
-
-    //end processing
-    b.set(true)
-
-    result.get
-  }
-
-  property("link executes cleanup code on process") = secure {
-    val adp = new AtomicBoolean(false)
-    val S = Strategy.Executor(Monitoring.defaultPool)
-    val hook = scalaz.stream.async.signalOf[Unit](())(S)
-    val other: Process[Task, Unit] = scalaz.stream.io.resource(
-      Task.delay(())
-    )(
-      _ => Task.delay{adp.set(true);println("cleaning up");()}
-    )(_ => Task.delay(()))
-    Monitoring.default.link(hook)(other)
-    hook.close.run
-    adp.get
-  }
-   */
 }
